+++ src/Jackalope/Version/VersionHistory.php
@@ -3,6 +3,9 @@
 namespace Jackalope\Version;
 
 use ArrayIterator;
+use Jackalope\Property;
+use PHPCR\NodeInterface;
+use PHPCR\Util\NodeHelper;
 use PHPCR\Version\VersionHistoryInterface;
 use PHPCR\Version\VersionInterface;
 use PHPCR\Version\VersionException;
@@ -33,6 +36,12 @@ class VersionHistory extends Node implements VersionHistoryInterface
      * @var array
      */
     protected $linearVersions = null;
+
+    /**
+     * Cache of the version labels.
+     * @var array
+     */
+    protected $versionLabels = null;
 
     /**
      * {@inheritDoc}
@@ -166,9 +175,11 @@ class VersionHistory extends Node implements VersionHistoryInterface
      */
     public function getVersionByLabel($label)
     {
-        // @codeCoverageIgnoreStart
-        throw new NotImplementedException();
-        // @codeCoverageIgnoreEnd
+        if (!$this->hasVersionLabel($label)) {
+            throw new VersionException("No label '$label'");
+        }
+
+        return $this->versionLabels[$label];
     }
 
     /**
@@ -178,9 +189,12 @@ class VersionHistory extends Node implements VersionHistoryInterface
      */
     public function addVersionLabel($versionName, $label, $moveLabel)
     {
-        // @codeCoverageIgnoreStart
-        throw new NotImplementedException();
-        // @codeCoverageIgnoreEnd
+        $this->initVersionLabels();
+        $version = $this->getVersion($versionName);
+        $path = $version->getPath();
+
+        $this->objectManager->addVersionLabel($path, $label, $moveLabel);
+        $this->versionLabels[$label] = $version;
     }
 
     /**
@@ -190,9 +204,13 @@ class VersionHistory extends Node implements VersionHistoryInterface
      */
     public function removeVersionLabel($label)
     {
-        // @codeCoverageIgnoreStart
-        throw new NotImplementedException();
-        // @codeCoverageIgnoreEnd
+        if (!$this->hasVersionLabel($label)) {
+            throw new VersionException("No label '$label'");
+        }
+
+        $version = $this->versionLabels[$label];
+        $this->objectManager->removeVersionLabel($version->getPath(), $label);
+        unset($this->versionLabels[$label]);
     }
 
     /**
@@ -202,9 +220,9 @@ class VersionHistory extends Node implements VersionHistoryInterface
      */
     public function hasVersionLabel($label, $version = null)
     {
-        // @codeCoverageIgnoreStart
-        throw new NotImplementedException();
-        // @codeCoverageIgnoreEnd
+        $labels = $this->getVersionLabels($version);
+
+        return in_array($label, $labels);
     }
 
     /**
@@ -214,9 +232,57 @@ class VersionHistory extends Node implements VersionHistoryInterface
      */
     public function getVersionLabels($version = null)
     {
-        // @codeCoverageIgnoreStart
-        throw new NotImplementedException();
-        // @codeCoverageIgnoreEnd
+        $this->initVersionLabels();
+        if ($version === null) {
+            return array_keys($this->versionLabels);
+        }
+
+        $versions = $this->getAllVersions();
+        $versionIsInHistory = false;
+
+        foreach ($versions as $versionCheck) {
+            /* @var VersionInterface $versionCheck */
+            if ($versionCheck->getIdentifier() == $version->getIdentifier()) {
+                $versionIsInHistory = true;
+                break;
+            }
+        }
+
+        if (!$versionIsInHistory) {
+            throw new VersionException(sprintf('Version %s not found in history of %s', $version->getIdentifier(), $this->getPath()));
+        }
+
+        $result = array();
+        foreach ($this->versionLabels as $label => $labelVersion) {
+            /* @var VersionInterface $labelVersion */
+            if ($labelVersion->getIdentifier() == $version->getIdentifier()) {
+                $result[] = $label;
+            }
+        }
+
+        return $result;
+    }
+
+    /**
+     * This method fetches all version labels, if the cache array is not initialized yet.
+     */
+    private function initVersionLabels()
+    {
+        if (!is_null($this->versionLabels)) {
+            return;
+        }
+
+        $this->versionLabels = array();
+        $node = $this->getNode('jcr:versionLabels');
+        foreach ($node->getProperties() as $property) {
+            /* @var Property $property */
+
+            if (NodeHelper::isSystemItem($node)) {
+                $name = $property->getName();
+                $value = $this->objectManager->getNodeByIdentifier($property->getValue()->getIdentifier(), 'Version\\Version');
+                $this->versionLabels[$name] = $value;
+            }
+        }
     }
 
     /**
@@ -247,5 +313,6 @@ class VersionHistory extends Node implements VersionHistoryInterface
     {
         $this->versions = null;
         $this->linearVersions = null;
+        $this->versionLabels = null;
     }
 }