+++ CHANGELOG.md
@@ -3,6 +3,6 @@ CHANGELOG
 
 dev-master
 ----------
-
+* feature #302 Added methods addVersionLabel and removeVersionLabel to VersioningInterface
 * bugfix #229 Userland node type filtering does not work
 * feature #245 Added NodeProcessor class which can be used by implementations to validate and process nodes.